% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelFitting.R
\name{analyse.casectrl}
\alias{analyse.casectrl}
\alias{analyse.casectrl.lrt}
\alias{analyse.quant}
\alias{analyse.casectrl.lrt}
\alias{analyse.quant}
\title{quantitative/case-control analysis}
\usage{
analyse.casectrl.lrt(obj, mode = "scaled", model = NULL, dnaDesign = NULL,
  rnaDesign = ~1)

analyse.quant(obj, mode = "quant", model = NULL, dnaDesign = NULL,
  rnaDesign = ~1)
}
\arguments{
\item{obj}{the MpraObject}

\item{model}{the model to fit}

\item{dnaDesign}{the design for the DNA model}

\item{rnaDesign}{the design for the RNA model}
}
\value{
the MpraObject with fitted models and condition test
}
\description{
test for difference between case and control enhancers or 
compare estimated magnitude against distribution of effect
on lower quantile.
}
