% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelFitting.R
\name{analyse.condition}
\alias{analyse.condition}
\alias{analyse.condition.lrt}
\alias{analyse.condition.ttest}
\alias{analyse.condition.lrt}
\alias{analyse.condition.ttest}
\title{differential activity analysis by condition}
\usage{
analyse.condition.lrt(obj, model = "gamma.pois", mode = NULL,
  dnaDesign = NULL, rnaDesign = NULL, condition_totest = NULL)

analyse.condition.ttest(obj, model = "gamma.pois", mode = "ttest",
  dnaDesign = NULL, rnaDesign = NULL, condition_totest = NULL)
}
\arguments{
\item{obj}{the MpraObject}

\item{model}{the model to use (TODO: if nul?)}

\item{mode}{options are "scaled" (default") or "full", see details}

\item{dnaDesign}{the design to use for the DNA model (see details)}

\item{rnaDesign}{the design to use for the RNA model (see details)}

\item{condition_totest}{the part of rnaDesign to test for (see details)}
}
\value{
the MpraObject with fitted models and condition test
}
\description{
test condition effect
}
\details{
TODO - explain input designs (either formula or model matrices)

TODO rna control usage

TODO - explain mode
}
