% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysisFunctions.R
\name{analyze.comparative.coef}
\alias{analyze.comparative.coef}
\title{Fit the model to enable coefficient-based testing (using the test.coef).
This method should be used when multiple conditions in the data are to be
tested using the same model.
For example, various stimulations compared with unstimulated.}
\usage{
analyze.comparative.coef(obj, dnaDesign, rnaDesign, use.controls = TRUE)
}
\arguments{
\item{obj}{the MpraObject}

\item{dnaDesign}{the design of the DNA model}

\item{rnaDesign}{the design of the RNA model}

\item{useControls}{if the experiment has negative controls, thes can be 
included in the model and used to correct for unwanted variation. Default is 
TRUE, and this is ignored if no controls exist in the experiment.}
}
\value{
the MpraObject with the fitted model, that can be passed to test.coef 
to test for significance of any coefficient in the model.
}
\description{
Fit the model to enable coefficient-based testing (using the test.coef).
This method should be used when multiple conditions in the data are to be
tested using the same model.
For example, various stimulations compared with unstimulated.
}
\note{
Currently, this mode only supports a single reference (intercept term).
}
