% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysisFunctions.R
\name{analyze.comparative.lrt}
\alias{analyze.comparative.lrt}
\title{Fit the model for a likelihood-ratio based testing of the significance of the
effect of the given condition.}
\usage{
analyze.comparative.lrt(obj, condition = NULL, mode = NULL,
  dnaDesign = NULL, rnaDesign = ~condition)
}
\arguments{
\item{obj}{the MPRAnalyze object}

\item{condition}{the condition to test. Must be a valid column name in the 
annotation of the object. Ignored in "full" mode.}

\item{mode}{the LRT mode to run in. See details}

\item{dnaDesign}{the design of the DNA model. See details}

\item{rnaDesign}{the design of the RNA model. See details}
}
\value{
the MPRAnalyze object, populated with fitted models
}
\description{
Fit the model for a likelihood-ratio based testing of the significance of the
effect of the given condition.
}
\details{
comparative LRT-based analysis can be run in three possible modes.
\itemize {
  \item "nocontrol": this is the default mode for scenarios where control 
  sequences (scrambles) are not available. In this mode, The full model folows
  the given design, and the reduced model is similar but with the condition
  factor removed.
  \item "scaled": this is the default mode if controls sequences are available.
  In this mode, a separate model is fitted to the control sequences, and the
  fitten parameters are incorporated in the data model to account for unknown
  sources of unqanted variation. The control model is used to scale the data 
  model parameters, similar to library depth factors.
  \item "full": this mode answers a slightly different statistical question.
  instead of measuring the effect of the condition explicitly, this mode test
  if the sequence at hand behaves differently from the control enhancers. In
  this mode a separate model is fit for the control enhancers, and incorporated
  fully into the data model for each sequence in th full model, or separately
  for the reduced model. This mode would detect any sequence with statistically
  significant deviance from the control distribution.
}
}
