% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysisFunctions.R
\name{analyze.lrt}
\alias{analyze.lrt}
\title{run LRT-based analysis}
\usage{
analyze.lrt(obj, condition = NULL, dnaDesign = NULL,
  rnaDesign = ~condition)
}
\arguments{
\item{obj}{the MPRAnalyze object}

\item{condition}{the condition to test. Must be a valid column name in the 
annotation of the object. Ignored in "full" mode.}

\item{dnaDesign}{the design of the DNA model. See details}

\item{rnaDesign}{the design of the RNA model. See details}

\item{mode}{the LRT mode to run in. See details}
}
\value{
the MPRAnalyze object, populated with fitted models
}
\description{
run LRT-based analysis
}
