% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysisFunctions.R
\name{analyze.quantitative.empirical}
\alias{analyze.quantitative.empirical}
\title{Fit the model for quantitative analysis}
\usage{
analyze.quantitative.empirical(obj)
}
\arguments{
\item{obj}{the MpraObject}
}
\value{
the MpraObject with fitted models
}
\description{
Fit the model for quantitative analysis
}
