% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/srcMPRAnalyze_doHypothesisTests.R
\name{calcTTest}
\alias{calcTTest}
\title{T-test on RNA/DNA ratios of case enhancer sequence versus control set
One-sided test for case ration being larger than control.
Appropriate to compare a set of ratios of a given enhancer with a set of control ratios,
i.e. multiple observations per enhancer.}
\usage{
calcTTest(obj, vecCaseIDs, vecCtrlIDs)
}
\description{
T-test on RNA/DNA ratios of case enhancer sequence versus control set
One-sided test for case ration being larger than control.
Appropriate to compare a set of ratios of a given enhancer with a set of control ratios,
i.e. multiple observations per enhancer.
}
\seealso{
calcZTest
}
\author{
David Sebastian Fischer
}
