% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/srcMPRAnalyze_doHypothesisTests.R
\name{calcZTest}
\alias{calcZTest}
\title{Z-score on ratios of cumulative RNA/DNA counts per case enhancer sequence and control set.
One-sided test for case ration being larger than control.
Appropriate to compare one ration of a given enhancer with a set of control ratios,
i.e. one observation per enhancer.}
\usage{
calcZTest(obj, vecCaseIDs, vecCtrlIDs)
}
\description{
Z-score on ratios of cumulative RNA/DNA counts per case enhancer sequence and control set.
One-sided test for case ration being larger than control.
Appropriate to compare one ration of a given enhancer with a set of control ratios,
i.e. one observation per enhancer.
}
\seealso{
calcTTest
}
\author{
David Sebastian Fischer
}
