% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/designMatFunctions.R
\name{checkForIntercept}
\alias{checkForIntercept}
\title{Return TRUE iff the given design has an intercept term}
\usage{
checkForIntercept(design)
}
\arguments{
\item{design}{either a formula or a design matrix}
}
\value{
TRUE iff the design has an intercept term
}
\description{
Return TRUE iff the given design has an intercept term
}
