% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{colAnnot}
\alias{colAnnot}
\title{Column annotations for the chromosomal-episomal dataset}
\format{\describe{
\item{batch}{batch identifier, factor}
\item{condition}{condition identifier, factor. WT corresponds to chromosomal
and MT corresponds to episomal}
\item{barcode}{barcode identifier, factor}
}}
\source{
\url{https://www.ncbi.nlm.nih.gov/pmc/articles/PMC5204343/}
}
\usage{
colAnnot
}
\description{
Column annotations for the chromosomal-episomal dataset
}
\keyword{datasets}
