% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/libraryDepth.R
\name{compute.depth}
\alias{compute.depth}
\title{compute depth factor}
\usage{
compute.depth(data, lib.factor, func)
}
\arguments{
\item{data}{the data matrix to compute the depth factor for}

\item{lib.factor}{the partitioning of the data matrix columns into library.
A separate depth factor is computed per library.}

\item{func}{the function to use to compute the library depth}
}
\value{
a numeric of length NCOL(data) with the appropriate epth factor for
each column (these are not unique values)
}
\description{
compute depth factor
}
