% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/srcMPRAnalyze_estimateDepthFactors.R
\name{computeSizeFactors}
\alias{computeSizeFactors}
\title{Compute a size factor for each sample}
\usage{
computeSizeFactors(matCountData)
}
\arguments{
\item{matCountData}{(matrix genes x samples)
Read count data.}
}
\value{
vecSizeFactors (numeric vector number of samples) 
   Model scaling factors for each sample which take
   sequencing depth into account (size factors).
}
\description{
Compute a size factor for each sample
}
\seealso{
Called by \link{computeNormConst}.
}
\author{
David Sebastian Fischer
}
