% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{control}
\alias{control}
\title{indices of control enhancers in the chromosomal-episomal dataset}
\format{logical vector, TRUE means the corresponding row in the observation
matrices are control features}
\source{
#' DNA observations for the chromosomal-episomal dataset

\url{https://www.ncbi.nlm.nih.gov/pmc/articles/PMC5204343/}
}
\usage{
control
}
\description{
indices of control enhancers in the chromosomal-episomal dataset
}
\keyword{datasets}
