% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/designMatFunctions.R
\name{correctForIDR}
\alias{correctForIDR}
\title{Correct RNA design matrix for quantitative IDR-based analyis}
\usage{
correctForIDR(obj, rnaDesign)
}
\description{
For IDR analysis we need a library-specific slope estimate. Since the library
isn't necessarily included in the design by the user, this function is used
to make sure that the library id is included in the design as the first term,
without an intercept.
}
