% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objectiveFunctions.R
\name{cost.model}
\alias{cost.model}
\title{likelihood wrapper to compute terms of likelihood with non-zero derivative
with respect to case dna model and rna model}
\arguments{
\item{theta}{vector of model parameters to evaluate likelihood for
(numeric, parameters)}

\item{theta.d}{vector of dna model parameters to condition likelihood on
(numeric, dna parameters)}

\item{theta.r}{vector of rna model parameters to condition likelihood on
(numeric, rna parameters)}

\item{llfnDNA}{cost function to compute dna likelihood terms on (function)}

\item{llfnRNA}{cost function to compute rna likelihood terms on (function)}

\item{dcounts}{the observed case DNA counts (integer, enhancers x samples)}

\item{rcounts}{the observed RNA counts (integer, enhancers x samples)}

\item{log.ddepth}{dna library size correction vector, log scale (numeric, samples)}

\item{log.rdepth}{rna library size correction vector, log scale (numeric, samples)}

\item{ddesign.mat}{the dna model design matrix (logical, samples x dna parameters)}

\item{rdesign.mat}{the rna model design matrix (logical, samples x rna parameters)}

\item{rdesign.ctrl.mat}{the control rna model design matrix 
(logical, samples x rna parameters)}
}
\description{
nomenclature cost.[model components to be estimated].[if ctrl obs are included in fitting]
}
