% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/libraryDepth.R
\docType{data}
\name{DEPTH_EST_FUNCTIONS}
\alias{DEPTH_EST_FUNCTIONS}
\title{easy access container for depth estimation functions}
\format{An object of class \code{list} of length 2.}
\usage{
DEPTH_EST_FUNCTIONS
}
\description{
easy access container for depth estimation functions
}
\keyword{datasets}
