% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hypothesisTesting.R
\name{est.mode}
\alias{est.mode}
\title{estimate mode of a sample}
\usage{
est.mode(x)
}
\arguments{
\item{x}{the sample to estimate the mode of}
}
\value{
the estimate mode of the sample distribution
}
\description{
estimate mode of a sample
}
\details{
the etimated mode is the center of the window that contains the most
observations. Window size is set to 1% of the range of values.
}
