% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/srcMPRAnalyze_estimateDepthFactors.R
\name{estimateDepthFactors}
\alias{estimateDepthFactors}
\title{Set depth factors in objectMPRAnalyze object}
\usage{
estimateDepthFactors(objectMPRAnalyze)
}
\description{
Set depth factors in objectMPRAnalyze object
}
\author{
David Sebastian Fischer
}
