% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/srcMPRAnalyze_estimateDispersions.R
\name{estimateDispersion}
\alias{estimateDispersion}
\title{Estimate dispersion parameters on MPRA data set}
\usage{
estimateDispersion(objMPRAnalyze, vecModelFactors, boolVerbose = TRUE)
}
\value{
(data frame genes x characteristics)
Table containing details of dispersion fit per gene.
}
\description{
Uses DESeq2 model and package components to estimate
a MAP (trend-smoothed) dispersion estimator for 
the RNA distribution per gene. The intial 
MLE is fitted by this package according to MPRA 
characteristics.
}
\author{
David Sebastian Fischer
}
