% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/libraryDepth.R
\name{estimateFactors}
\alias{estimateFactors}
\title{estimate the library depth factors}
\usage{
estimateFactors(counts, annotations, lib.factor = NULL,
  depth.estimator = "uq")
}
\arguments{
\item{counts}{the counts}

\item{annotations}{the annotations (data.frame)}

\item{lib.factor}{the name of the factors describing the libraries}

\item{depth.estimator}{a character indicating which depth estimation to use.
Currently supported values are "uq" for upper quantile (default) and "rle"
for RLE (uses geometric mean, and is therefore not recommended if libraries
have 0 counts)}
}
\value{
computed library depth corection factors
}
\description{
estimate the library depth factors
}
