% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/randomEffectVariance.R
\name{estimateRandomEffectVariance}
\alias{estimateRandomEffectVariance}
\title{estimate random effect variance}
\usage{
estimateRandomEffectVariance(obj, rand.factor, lib.factor, use_vst = F)
}
\arguments{
\item{obj}{the MpraObject}

\item{rand.factor}{the factor on which to estimate random effect}

\item{lib.factor}{the factor associating each sample to a library. Can be a
factor or the name of a column in the object's colAnnot. If not provided, the
data is assumed to have been generated from a single library, and constant
library depth is set.}
}
\value{
the MpraObject with estimated value for random effect variance
}
\description{
estimate random effect variance
}
\examples{
data <- simulateMPRA(tr = rep(2,10), da=NULL, nbatch=2, nbc=20)
obj <- MpraObject(dnaCounts = data$obs.dna,
                  rnaCounts = data$obs.rna,
                  colAnnot = data$annot)
obj <- estimateRandomEffectVariance(obj, lib.factor = "sample", rand.factor = "barcode")
}
