% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelSelection.R
\name{evalModels}
\alias{evalModels}
\title{evaluate the performance of the supported models on the given data}
\usage{
evalModels(obj)
}
\arguments{
\item{obj}{the MpraObject}
}
\description{
evaluate the performance of the supported models on the given data
}
