% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimizers.R
\name{fit.dna.controlrna}
\alias{fit.dna.controlrna}
\title{Fit the DNA model conditioned on an RNA model}
\usage{
fit.dna.controlrna(model, dcounts, rcounts, r.coef, ddepth, rdepth, ddesign.mat,
  rdesign.mat, d2rdesign.mat)
}
\arguments{
\item{model}{noise model}

\item{dcounts}{the DNA counts}

\item{rcounts}{the RNA counts}

\item{ddepth}{dna library size correction vector (numeric, dna samples)}

\item{rdepth}{rna library size correction vector (numeric, rna samples)}

\item{ddesign.mat}{the dna model design matrix (logical, rna samples x dna parameters)}

\item{rdesign.mat}{the rna model design matrix (logical, rna samples x rna parameters)}

\item{d2rdesign.mat}{the transition matrix relating DNA estimates to RNA observations (logical, rna sample x dna parameters)}
}
\description{
This is used to optimize an enhancer-based model for each enhancer while
assuming that the RNA model is identical to the control model. This model
can be used as a reduced model in LRT settings.
}
