% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimizers.R
\name{fit.dnarna.onlyctrl.iter}
\alias{fit.dnarna.onlyctrl.iter}
\title{fit the control-based joint model. Seperate DNA model per enhancer, joint RNA
model.}
\usage{
fit.dnarna.onlyctrl.iter(model, dcounts, rcounts, ddepth, rdepth,
  ddesign.mat, rdesign.mat, d2rdesign.mat, BPPARAM,
  print.progress = TRUE)
}
\arguments{
\item{model}{noise model}

\item{dcounts}{the DNA counts}

\item{rcounts}{the RNA counts}

\item{ddepth}{dna library size correction vector (numeric, dna samples)}

\item{rdepth}{rna library size correction vector (numeric, rna samples)}

\item{ddesign.mat}{the dna model design matrix 
(logical, rna samples x dna parameters)}

\item{rdesign.mat}{the rna model design matrix 
(logical, rna samples x rna parameters)}

\item{d2rdesign.mat}{the transition matrix relating DNA estimates to RNA 
observations (logical, rna sample x dna parameters)}

\item{BPPARAM}{the parallelization backend to use}

\item{print.progress}{print a progress report on the fitting (default:TRUE),
recommended since the joint model fitting is computationally intensive}
}
\value{
a list with components:
\itemize{
    \item d.fitval: the fitted values of the DNA counts
    \item d.est: the estimated true DNA levels (corrected for library size)
    \item d.coef: the fitted mode parameters for the DNA counts
    \item d.se: the standard errors of the estimates of the DNA counts
    \item r.est: the fitted values of the RNA counts
    \item r.fitval: the fitted mode parameters for the RNA counts
    \item r.est: the estimated true RNA levels (corrected for library size)
    \item r.se: the standard errors of the estimates of the RNA counts
    \item ll: the log likelihood of the model
}
}
\description{
fit the control-based joint model. Seperate DNA model per enhancer, joint RNA
model.
}
