% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimizers.R
\name{fit.dnarna}
\alias{fit.dnarna}
\alias{fit.dnarna.noctrlobs}
\alias{fit.dnarna.wctrlobs.iter}
\alias{fit.dnarna.noctrlobs}
\alias{fit.dnarna.wctrlobs.iter}
\title{Fit dna and rna model to a given enhancer}
\usage{
fit.dnarna.noctrlobs(model, dcounts, rcounts, ddepth, rdepth, rctrlscale,
  ddesign.mat, rdesign.mat, d2rdesign.mat, rdesign.ctrl.mat,
  theta.d.ctrl.prefit, compute.hessian)

fit.dnarna.wctrlobs.iter(model, dcounts, rcounts, ddepth, rdepth,
  rctrlscale = NULL, ddesign.mat, rdesign.mat, d2rdesign.mat,
  rdesign.ctrl.mat, theta.d.ctrl.prefit, compute.hessian)
}
\arguments{
\item{model}{noise model}

\item{dcounts}{the DNA counts}

\item{rcounts}{the RNA counts}

\item{ddepth}{dna library size correction vector (numeric, dna samples)}

\item{rdepth}{rna library size correction vector (numeric, rna samples)}

\item{rctrlscale}{control-based correction scalers}

\item{ddesign.mat}{the dna model design matrix (logical, rna samples x dna parameters)}

\item{rdesign.mat}{the rna model design matrix (logical, rna samples x rna parameters)}

\item{d2rdesign.mat}{the transition matrix relating DNA estimates to RNA observations (logical, rna sample x dna parameters)}

\item{rdesign.ctrl.mat}{the control rna model design matrix 
(logical, samples x rna parameters)}

\item{theta.d.ctrl.prefit}{ctrl dna model parameters to condition likelihood on
(numeric, control enhancers x dna parameters)}

\item{compute.hessian}{if TRUE (default), compute the Hessian matrix of the
coefficients to facilitate coefficient-based hypothesis testing}
}
\value{
a list with components:
\itemize{
    \item d.fitval: the fitted values of the DNA counts
    \item d.est: the estimated true DNA levels (corrected for library size)
    \item d.coef: the fitted mode parameters for the DNA counts
    \item d.se: the standard errors of the estimates of the DNA counts
    \item r.est: the fitted values of the RNA counts
    \item r.fitval: the fitted mode parameters for the RNA counts
    \item r.est: the estimated true RNA levels (corrected for library size)
    \item r.se: the standard errors of the estimates of the RNA counts
    \item ll: the log likelihood of the model
}
}
\description{
Optim wrapper that performs numerical optimisation and extracts results.
Depending on the function chosen, a single optimisation or iterative estimation
are performed.
}
\details{
dna model
The dna model is encoded as a vector where the first parameter
is the variance-link (ln: stdv, nb: dispersion) and the remaining parameters
are the coefficient of the linear model that consitutes the mean parameter,
as encoded in the model matrix.

rna model
The case rna model and the control rna interaction terms are encoded as seperate
linear models so that they can be added up by design matrix concatenation (cbind)
if a control enhancer is considered. Note that the first parameter
may be a variance link parameter (ln.nb: nb: dispersion) which does not contribute
to the linear model that constitutes the mean parameter. As all parameters only act
on the mean parameter in the case of gamma.pois, the rna model matrix is 
augmented by a first column containing only padding 0s if such a variance parameter
exists, so that the mean parameter can be computed via the same matrix multiplication
irrespective of whether the rna model contains such parameter that does not contribute
to the mean model. The code that throws away non-used coefficients before fitting
is blocked for this parameter. Note that the ctrl rna model is not allowed to have
and additional variance coefficient so that the testing is only on the first moment.
This is guaranteed by the default that this non-use cofficient is discarded.
}
