% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelFitting.R
\name{fit.gammaDNA}
\alias{fit.gammaDNA}
\title{fit a Gamma GLM to the DNA count data}
\usage{
fit.gammaDNA(dcounts, depth, design.mat, control)
}
\arguments{
\item{dcounts}{the DNA counts (integer, N)}

\item{depth}{library size correction factors (numeric, N)}

\item{design.mat}{the design matrix (logical, N x K)}

\item{control}{list of parameter to be passed to `optim` as
the control argument. NOTE: since this is a maximization problem,
control$fnscale should be set to a negative number (idealy -1)}
}
\value{
a list with components:
\itemize {
    \item est: the fitted values
    \item model: the fitted coefficients for the parameters
    \item se: the standard errors of the coefficients
    \item ll: the log likelihood of the model
}
}
\description{
fit a Gamma GLM to the DNA count data
}
