% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelFitting.R
\name{fit.mixture.gammaPoisson}
\alias{fit.mixture.gammaPoisson}
\title{Fit a gamma-poisson mixture model to the DNA-RNA counts. A gamma distribution
is fitted to the DNA counts and a Poisson noise is added for the RNA counts
which results in a Negative Binomial distribution for the RNA. The model is
fit simultaneously.}
\usage{
fit.mixture.gammaPoisson(dcounts, rcounts, ddepth, rdepth, ddesign.mat,
  rdesign.mat, control)
}
\arguments{
\item{dcounts}{the DNA counts}

\item{rcounts}{the RNA counts}

\item{ddepth}{DNA library size correction factors}

\item{rdepth}{RNA library size correction factors}

\item{ddesign.mat}{the design matrix for the DNA}

\item{rdesign.mat}{the design matrix for the RNA}

\item{control}{control arguments to pass to `optim`. NOTE: since this is a
maximization problem, control$fnscale should be set to a negative number.}
}
\value{
a list with components:
\itemize {
    \item d.est: the fitted values of the DNA counts
    \item d.model: the fitted mode parameters for the DNA counts
    \item d.se: the standard errors of the estimates of the DNA counts
    \item r.est: the fitted values of the RNA counts
    \item r.model: the fitted mode parameters for the RNA counts
    \item r.se: the standard errors of the estimates of the RNA counts
    \item ll: the log likelihood of the model
}
}
\description{
Fit a gamma-poisson mixture model to the DNA-RNA counts. A gamma distribution
is fitted to the DNA counts and a Poisson noise is added for the RNA counts
which results in a Negative Binomial distribution for the RNA. The model is
fit simultaneously.
}
