% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelFitting.R
\name{fit.nbRNA}
\alias{fit.nbRNA}
\title{fit a negative binomial model to the RNA counts using a point estimator for
the DNA counts}
\usage{
fit.nbRNA(rcounts, depth, d.est, design.mat, control)
}
\arguments{
\item{rcounts}{RNA counts}

\item{depth}{library size correction factors}

\item{d.est}{point estimates of the DNA base counts}

\item{design.mat}{the RNA design matrix}

\item{control}{optimization parameters to be passed to \code{optim}. NOTE:
since this is a maximization problem, control$fnscale should be set to a
negative number (idealy -1)}
}
\value{
a list with components:
\itemize {
    \item est: the fitted values
    \item model: the fitted coefficients for the parameters
    \item se: the standard errors of the coefficients
    \item ll: the log likelihood of the model
}
}
\description{
fit a negative binomial model to the RNA counts using a point estimator for
the DNA counts
}
