% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/srcMPRAnalyze_fitModels.R
\name{fitModels}
\alias{fitModels}
\title{Fits constant models to an MPRA dataset}
\usage{
fitModels(obj, vecModelFacRNA, vecModelFacRNACtrl, vecModelFacDNA,
  MAXIT = 1000, RELTOL = 10^(-8), boolVerbose = TRUE)
}
\arguments{
\item{obj}{(object class obj)
Object to be fit.}

\item{vecModelFac}{(vector of strings number of discrete model variables)}
}
\description{
[Model fitting function hierarchy: 1 out of 4]
This primary wrapper coordinates fitting of impulse and constant model
to separate conditions according to the differential expression
mode (case-only or case-control).
}
\seealso{
Calls \link{fitConstImpulse}
once for each condition with the appropriate parameters and samples.
}
\author{
David Sebastian Fischer
}
