% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/srcMPRAnalyze_fitModels.R
\name{fitRNA_pointDNAnbRNA}
\alias{fitRNA_pointDNAnbRNA}
\title{Fit negative binomial RNA model based on DNA point estimator}
\usage{
fitRNA_pointDNAnbRNA(matRNACounts, matDNAEst, scaDisp, vecRNADepth,
  lsvecidxBatchRNA, boolBaselineCtrl = FALSE, lsvecidxBatchRNACtrl = NULL,
  MAXIT = 1000, RELTOL = 10^(-8))
}
\arguments{
\item{scaDisp}{(scalar) Gene-wise 
negative binomial dispersion hyper-parameter.}

\item{MAXIT}{(scalar) [Default 1000] 
Maximum number of BFGS iterations for model fitting with \link{optim}.}

\item{RELTOL}{(scalar) [Default 10^(-8)]
Maximum relative change in loglikelihood to reach convergence in
numerical optimisation by BFGS in \link{optim}.}

\item{vecCounts}{(numeric vector number of samples)
Read count data.}

\item{vecSizeFactors}{(numeric vector number of samples) 
Model scaling factors for each sample which take
sequencing depth into account (size factors).}

\item{lsvecidxBatch}{(list length number of confounding variables)
List of index vectors. 
One vector per confounding variable.
Each vector has one entry per sample with the index batch
within the given confounding variable of the given sample.
Batches are enumerated from 1 to number of batches.}

\item{trace}{(scalar) [Defaul 0]
Reporting parameter of \link{optim}.}

\item{REPORT}{(scalar) [Default 10]
Reporting parameter of \link{optim}.}
}
\value{
(list) List of constant fit parameters and results.
   \itemize{
     \item scaMu (scalar) Maximum likelihood estimator of
     negative binomial mean parameter.
     \item lsvecBatchFactors (list length number of confounders)
     List of vectors of scalar batch correction factors for each sample.
     These are also maximum likelihood estimators.
     NULL if no confounders given.
     \item scaDispParam (scalar) Dispersion parameter estimate
     used in fitting (hyper-parameter).
     \item scaLL (scalar) Loglikelihood of data under maximum likelihood
     estimator model.
     \item scaConvergence (scalar) 
     Convergence status of optim on constant model.
   }
}
\description{
With or without dispersion.
}
\seealso{
Called by \link{fitConstImpulseGene} to fit constant
   model to samples of one condition and one gene.
   Calls constant model cost function 
   \link{evalLogLik} within \link{optim}.
}
\author{
David Sebastian Fischer
}
