% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/likelihoodFunctions.R
\name{get.ll.functions}
\alias{get.ll.functions}
\title{Get appropriate likelihood function accourding to the selected model}
\usage{
get.ll.functions(model)
}
\arguments{
\item{model}{the distributional model selected}
}
\value{
a list with two items: dna, rna. Each item is the corresponding 
likelihod function
}
\description{
Get appropriate likelihood function accourding to the selected model
}
