% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilFunctions.R
\name{getAlpha}
\alias{getAlpha}
\title{return the fitted value for alpha (transcription rate).}
\usage{
getAlpha(obj, term = NULL, value = NULL, full = TRUE)
}
\arguments{
\item{obj}{the MpraObject}

\item{term}{the term to get the alpha of (see details)}

\item{value}{the value of the term to get the alpha of (Se details)}

\item{full}{if true, return alpha of the full model (default), otherwise of
the reduced model (only applies if an LRT-based analysis was used)}
}
\description{
return the fitted value for alpha (transcription rate).
}
\details{
return the estimate for transcription rate as fitted by the package.
If the design is intercepted, then by default the baseline (intercept) rate is
returned. Otherwise, term and value must be provided, such that term is a 
valid term in the design provided to the fit, and value is one of the 
levels in the term.
}
