% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/design.R
\name{getDesign}
\alias{getDesign}
\title{one condition scenario:
- normalize for library size
- choose model (?)
- fit model (single model!)
- extract model params
- visualizations}
\usage{
getDesign(obj, formula)
}
\description{
Two conditions scenario:
- normalize for library size
- choose model
- fit model (full + reduced!)
- differential activity analysis
- Visualizations
}
