% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelFitting.R
\name{getDesignMat}
\alias{getDesignMat}
\title{get a design matrix from the input design}
\usage{
getDesignMat(obj, design, mode)
}
\arguments{
\item{obj}{the MpraObject}

\item{design}{the input design. A matrix is returned as is, a formula is
expanded to a model matrix using the object colAnnot. If design is NULL, an
intercept-only design matrix is created and returned}

\item{testcondition}{condition to substract from formulae TODO this}
}
\value{
a design matrix

##TODO: this should take the mode into account
}
\description{
get a design matrix from the input design
}
