% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mpraObject.R
\name{getDNAFits}
\alias{getDNAFits}
\title{Get DNA model fits from an MpraObject}
\usage{
getDNAFits(obj, enhancers = NULL, depth = FALSE, full = TRUE)
}
\arguments{
\item{obj}{MpraObject to extract from}

\item{enhancers}{enhancer to extract}

\item{depth}{include depth correction}

\item{full}{whether to extract from full model}
}
\value{
DNA fits (numeric, enhancers x samples)
}
\description{
Get DNA model fits from an MpraObject
}
