% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessFunctions.R
\name{getFits.DNA}
\alias{getFits.DNA}
\title{Get DNA model fits from an MpraObject}
\usage{
getFits.DNA(obj, enhancers = NULL, depth = TRUE, full = TRUE,
  transition = FALSE)
}
\arguments{
\item{obj}{MpraObject to extract from}

\item{enhancers}{enhancer to extract}

\item{depth}{include depth correction}

\item{full}{whether to extract from full model}

\item{transition}{use the DNA->RNA transition matrix (deafult: FALSE)}
}
\value{
DNA fits (numeric, enhancers x samples)
}
\description{
Get DNA model fits from an MpraObject
}
