% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessFunctions.R
\name{getLibraryDepthFactors}
\alias{getLibraryDepthFactors}
\title{extract library depth factors.}
\usage{
getLibraryDepthFactors(obj)
}
\arguments{
\item{obj}{the object o extract the factors from}
}
\value{
a list with two elements: 'dna' with the dna library depth factors, 
and 'rna' with the rna factors
}
\description{
extract library depth factors.
}
