% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mpraObject.R
\name{getResults}
\alias{getResults}
\title{Get the analysis results after running an analysis function}
\usage{
getResults(obj)
}
\arguments{
\item{obj}{the MpraObject}
}
\value{
a fata.frame with the analysis results. Columns vary according to
the type of analysis
}
\description{
Get the analysis results after running an analysis function
}
