% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilFunctions.R
\name{isNestedDesign}
\alias{isNestedDesign}
\title{Return TRUE iff the reduced design is nested within the full design}
\usage{
isNestedDesign(full, reduced)
}
\arguments{
\item{full}{the full design (formula)}

\item{reduced}{the reduced design (formula)}
}
\value{
TRUE iff the reduced design is nested in the full design
}
\description{
Return TRUE iff the reduced design is nested within the full design
}
