% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/srcMPRAnalyze_classMPRAnalyzeObject.R
\docType{methods}
\name{list_accession}
\alias{list_accession}
\alias{names.MPRAnalyzeObject}
\alias{names,MPRAnalyzeObject-method}
\alias{$,MPRAnalyzeObject-method}
\alias{[[,MPRAnalyzeObject,character,missing-method}
\alias{list_accession}
\alias{list_accession}
\alias{list_accession}
\title{List-like accessor methods for MPRAnalyze}
\usage{
\S4method{names}{MPRAnalyzeObject}(x)

\S4method{[[}{MPRAnalyzeObject,character,missing}(x, i, j, ...)

\S4method{$}{MPRAnalyzeObject}(x, name)
}
\arguments{
\item{x}{(MPRAnalyzeObject) MPRAnalyze output object.}

\item{i, name}{(idx or str) Name or index of core output element of MPRAnalyzeObject}

\item{j}{Not used, only vectors.}

\item{...}{Not used.}
}
\value{
Names of core output in MPRAnalyzeObject.

Target element from MPRAnalyzeObject

Target element from MPRAnalyzeObject
}
\description{
Allow usage of MPRAnalyze ouput object like a list with
respect to the main output: MPRAnalyze
}
\author{
David Sebastian Fischer
}
