% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/likelihoodFunctions.R
\name{ll.dna}
\alias{ll.dna}
\alias{ll.dna.gamma.pois}
\alias{ll.dna.ln.nb}
\alias{ll.dna.gamma.pois}
\alias{ll.dna.ln.nb}
\title{likelihood of dna observations}
\usage{
ll.dna.gamma.pois(theta, dcounts, log.ddepth, ddesign.mat)

ll.dna.ln.nb(theta, dcounts, log.ddepth, ddesign.mat)
}
\arguments{
\item{theta}{the vector of rna model parameters to evaluate likelihood for
(numeric, rna parameters)}

\item{dcounts}{the observed DNA counts (integer, enhancers x samples)}

\item{log.ddepth}{dna library size correction vector, log scale (numeric, samples)}

\item{ddesign.mat}{the dna model design matrix (logical, samples x dna parameters)}
}
\value{
negative log likelihood of dna  observations under the specified model
}
\description{
likelihood of dna observations
}
\details{
matrix multiplications: object dimensions
theta: (dna param)
ddesign.mat: (samples x dna param-1)
log.ddepth, dcounts: (samples)

matrix multiplications: d.est
d.est <- theta[1] + ddesign.mat %*% theta[-1] + log.ddepth (gamma.pois)
d.est <- ddesign.mat %*% theta[-1] + log.ddepth (ln.nb)
dimensions( ddesign.mat %*% theta[-1] )
=(samples x dna param-1) x (dna param -1)
}
