% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objectiveFunctions.R
\name{ll.gammaDNA}
\alias{ll.gammaDNA}
\title{Likelihood objective function for gamma distribution for the DNA counts}
\usage{
ll.gammaDNA(theta, dcounts, depth, design.mat)
}
\arguments{
\item{theta}{a vector of parameters to be optimized. First is the shape
parameter (alpha), the rest correspond to columns of the design matrix.
Note that the first column is the inverse of beta, since that makes computing
the estimate easier (numeric, K + 1)}

\item{dcounts}{the DNA counts (integer, N)}

\item{depth}{library size correction factors (numeric, N)}

\item{design.mat}{the design matrix (logical, N x K)}
}
\value{
the log likelihood of the specified model
}
\description{
Likelihood objective function for gamma distribution for the DNA counts
}
