% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objectiveFunctions.R
\name{ll.lnDNA}
\alias{ll.lnDNA}
\title{likelihood objective function for lognormal point estimator for DNA counts}
\usage{
ll.lnDNA(theta, dcounts, depth, design.mat)
}
\arguments{
\item{theta}{the vector of parameters to fit (numeric, K + 1). First parameter
in the vector is the dispersion of the dstribution, the rest are ordered
in correspondence with the columns in the design matrix}

\item{dcounts}{the observed DNA counts (integer, N)}

\item{depth}{library size correction vector (numeric, N)}

\item{design.mat}{the design matrix (logical, N x K)}
}
\value{
the log likeihood of the observed counts assuming the model specified
by the parameters in theta and the esign matrix
}
\description{
likelihood objective function for lognormal point estimator for DNA counts
}
