% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objectiveFunctions.R
\name{ll.mixture.gammaPoisson}
\alias{ll.mixture.gammaPoisson}
\title{likelihood objective function for joint gamma-poisson mixture model}
\usage{
ll.mixture.gammaPoisson(theta, dcounts, rcounts, ddepth, rdepth, ddesign.mat,
  rdesign.mat)
}
\arguments{
\item{theta}{the parameter vector being optimized (numeric, K + M + 1)}

\item{dcounts}{the DNA count vector (integer, N)}

\item{rcounts}{the RNA count vector (integer, N)}

\item{ddepth}{the DNA library size correction factors (numeric, N)}

\item{rdepth}{the RNA library size correction factors (numeric, N)}

\item{ddesign.mat}{the DNA design matrix (logical, N x K)}

\item{rdesign.mat}{the RNA design matrix (logical, N x M)}
}
\description{
likelihood objective function for joint gamma-poisson mixture model
}
