% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objectiveFunctions.R
\name{ll.nbRNA}
\alias{ll.nbRNA}
\title{likelihood objective function for negative binomial RNA counts given a DNA
count estimator}
\usage{
ll.nbRNA(theta, rcounts, depth, d.est, design.mat)
}
\arguments{
\item{theta}{the parameter vector being optimized. First value is the
dispersion parameter, afterhich come factors ordered in correspondence
with the columns in the design.mat matrix (numeric, K + 1)}

\item{rcounts}{the RNA count vector (integer, N)}

\item{depth}{library size correction factors (numeric, N)}

\item{d.est}{the DNA point estimator to use, in log space (numeric, N)}

\item{design.mat}{the design matrix (logical, N x K)}
}
\value{
the log likelihood of the specified model
}
\description{
likelihood objective function for negative binomial RNA counts given a DNA
count estimator
}
