% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/likelihoodFunctions.R
\name{ll.rna}
\alias{ll.rna}
\alias{ll.rna.gamma.pois}
\alias{ll.rna.ln.nb}
\alias{ll.rna.ln.ln}
\title{likelihood of rna observations}
\usage{
ll.rna.gamma.pois(theta, theta.d, rcounts, log.rdepth, d2rdesign.mat,
  rdesign.mat)

ll.rna.ln.nb(theta, theta.d, rcounts, log.rdepth, d2rdesign.mat,
  rdesign.mat)

ll.rna.ln.ln(theta, theta.d, rcounts, log.rdepth, d2rdesign.mat,
  rdesign.mat)
}
\arguments{
\item{theta}{the vector of rna model parameters to evaluate likelihood for
(numeric, rna parameters)}

\item{theta.d}{dna model parameters to condition likelihood on}

\item{rcounts}{the observed RNA counts (integer, samples x enhancers)}

\item{log.rdepth}{rna library size correction vector, log scale (numeric, samples)}

\item{d2rdesign.mat}{the trasitional matrix to distribute and match DNA 
estimates to RNA observations (logical, rna samples x dna parameters)}

\item{rdesign.mat}{the rna model design matrix (logical, rna samples x rna parameters)}
}
\value{
negative log likelihood of rna observations under the specified model
}
\description{
likelihood of rna observations
}
