% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/srcMPRAnalyze_classMPRAnalyzeObject.R
\docType{class}
\name{MPRAnalyzeObject-class}
\alias{MPRAnalyzeObject-class}
\title{Container class for MPRAnalyze output}
\description{
MPRAnalyze output and intermediate results including as model fits.
}
\author{
David Sebastian Fischer
}
