% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MpraObject.R
\docType{class}
\name{MpraObject-class}
\alias{MpraObject-class}
\title{The main container class for MPRAnalyze}
\description{
The main container class for MPRAnalyze
}
\section{Slots}{

\describe{
\item{\code{dnaCounts}}{matrix of DNA counts}

\item{\code{rnaCounts}}{matrix of RNA counts}

\item{\code{dnaAnnot}}{DNA column annotations 
(info on condition, batch, barcode, etc)}

\item{\code{rnaAnnot}}{RNA column annotations 
(info on condition, batch, barcode, etc)}

\item{\code{controls}}{indices of negative controls}

\item{\code{controls.forfit}}{indices of negative controls used for fitting in `full`
mode}

\item{\code{lib.factor}}{a factor with a level for each library, used for depth 
estimation and in `idr` quantitative mode}

\item{\code{dnaDepth}}{library depth correction factors for DNA libraries}

\item{\code{rnaDepth}}{library depth correction factors for RNA libraries}

\item{\code{rnaCtrlScale}}{control-based correction factors for `scaled` modes}

\item{\code{mode}}{analysis mode (identifies the type of analysis performed)}

\item{\code{model}}{id of the distributional model used}

\item{\code{designs}}{a Designs object containing the various design matrices used}

\item{\code{modelFits}}{fitted models, populated by an analysis function}

\item{\code{modelFits.red}}{fitted reduced models, populated by an LRT analysis 
function}

\item{\code{modelPreFits.dna.ctrl}}{fitted models for control DNA}

\item{\code{BPPARAM}}{The BiocParallel parallelization backend to use throughout}
}}

