% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MpraObject.R
\docType{methods}
\name{dnaCounts}
\alias{dnaCounts}
\alias{dnaCounts,MpraObject-method}
\alias{rnaCounts}
\alias{rnaCounts,MpraObject-method}
\alias{dnaAnnot}
\alias{dnaAnnot,MpraObject-method}
\alias{rnaAnnot}
\alias{rnaAnnot,MpraObject-method}
\alias{controls}
\alias{controls,MpraObject-method}
\alias{dnaDepth}
\alias{dnaDepth,MpraObject-method}
\alias{rnaDepth}
\alias{rnaDepth,MpraObject-method}
\alias{model}
\alias{model,MpraObject-method}
\title{get properties from a given MpraObject}
\usage{
dnaCounts(obj)

\S4method{dnaCounts}{MpraObject}(obj)

rnaCounts(obj)

\S4method{rnaCounts}{MpraObject}(obj)

dnaAnnot(obj)

\S4method{dnaAnnot}{MpraObject}(obj)

rnaAnnot(obj)

\S4method{rnaAnnot}{MpraObject}(obj)

controls(obj)

\S4method{controls}{MpraObject}(obj)

dnaDepth(obj)

\S4method{dnaDepth}{MpraObject}(obj)

rnaDepth(obj)

\S4method{rnaDepth}{MpraObject}(obj)

model(obj)

\S4method{model}{MpraObject}(obj)
}
\arguments{
\item{obj}{the MpraObject to extract the property from}
}
\value{
The value of the property
}
\description{
get properties from a given MpraObject
}
\examples{
data <- simulateMPRA(tr = rep(2,15), nbatch=2, nbc=15)
obj <- MpraObject(dnaCounts = data$obs.dna, 
                  rnaCounts = data$obs.rna, 
                  colAnnot = data$annot,
                  controls = 1:5)
dnaCounts <- dnaCounts(obj)
rnaCounts <- rnaCounts(obj)
dnaAnnot <- dnaAnnot(obj)
rnaAnnot <- rnaAnnot(obj)
controls <- controls(obj)
dnaDepth <- dnaDepth(obj)
rnaDepth <- rnaDepth(obj)
model <- model(obj)
}
