% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotresults.R
\name{plot.volcano}
\alias{plot.volcano}
\title{Scatter plot observed log-fold changes and p-values for condition test:
volcano plot}
\usage{
\method{plot}{volcano}(obj)
}
\arguments{
\item{obj}{(MpraObject)
MpraObject from which enhancer is to be visualised.}
}
\value{
(ggplot) scatter plot
}
\description{
Scatter plot observed log-fold changes and p-values for condition test:
volcano plot
}
