% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotresults.R
\name{plotAlphaRatio}
\alias{plotAlphaRatio}
\title{create a scatter plot of the model alpha (trascription rate) vs. the naive
estimator (RNA / DNA).}
\usage{
plotAlphaRatio(obj, condition = NULL, logScale = TRUE)
}
\arguments{
\item{obj}{the MpraObject}

\item{condition}{if NULL (default) create a single plot of the model baseline
(intercept) vs the total data. Else, condition must be a valid factor in the 
object column annotations that was included in the design. In that case, a 
plot is generated for each level of the factor}

\item{logScale}{iff TRUE, figure is plotted in log scale}
}
\description{
create a scatter plot of the model alpha (trascription rate) vs. the naive
estimator (RNA / DNA).
}
