% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plottingFunctions.R
\name{plotMeanVariance}
\alias{plotMeanVariance}
\title{plot the relationship between the mean and the variance of the DNA and RNA 
distributions. Plots are in log scale.}
\usage{
plotMeanVariance(obj)
}
\arguments{
\item{obj}{the MpraObject to plot}
}
\value{
a list of two ggplot objects: 'dna' and 'rna'.
}
\description{
plot the relationship between the mean and the variance of the DNA and RNA 
distributions. Plots are in log scale.
}
