% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelSelection.R
\name{plotModelPerformance}
\alias{plotModelPerformance}
\title{plot various diagnostic plots to inform the user on model performance and
inform manual model selection}
\usage{
plotModelPerformance(obj)
}
\arguments{
\item{obj}{the MpraObject}
}
\description{
plot various diagnostic plots to inform the user on model performance and
inform manual model selection
}
