% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plottingFunctions.R
\name{plotObsExpDistributions}
\alias{plotObsExpDistributions}
\title{Plot the observed and expected distributions of a given enhancer}
\usage{
plotObsExpDistributions(obj, enhancer, bins = NULL)
}
\arguments{
\item{obj}{MpraObject after model fitting}

\item{enhancer}{the id of the enhancer to plot (index or name)}

\item{rna}{if TRUE, plot the RNA distribution. Otherwise plot the DNA.}
}
\description{
Plot the observed and expected distributions of a given enhancer
}
