% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotresults.R
\name{plotPvalCDF}
\alias{plotPvalCDF}
\title{plot the CDF of the pvalues}
\usage{
plotPvalCDF(p, categories, adjusted = FALSE)
}
\arguments{
\item{p}{the p-values to plot}

\item{categories}{a factor (same length as p) dividing the p-values into 
categories to be plotted separately}

\item{adjusted}{if TRUE, adjust the p-values using the BH FDR method.}
}
\description{
plot the CDF of the pvalues
}
