% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilFunctions.R
\name{reformatModels}
\alias{reformatModels}
\title{Reformat a list of models (return values of a fit.* function) to a list
of model parameters containing the corresponding parameters, for easier 
access and control}
\usage{
reformatModels(models)
}
\arguments{
\item{models}{the models to reformet, should be a list of results from a 
fit.* function}
}
\value{
a list of formatted extacted properties
}
\description{
Reformat a list of models (return values of a fit.* function) to a list
of model parameters containing the corresponding parameters, for easier 
access and control
}
