% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilFunctions.R
\name{resampleObs}
\alias{resampleObs}
\title{Resample observations of enhancer from fit distribution}
\usage{
resampleObs(obj, enhancer = NULL, full = TRUE)
}
\arguments{
\item{obj}{MpraObject to extract from}

\item{enhancer}{enhancer to extract}

\item{full}{whether to extract from full model}
}
\value{
resampled observations
}
\description{
Resample observations of enhancer from fit distribution
}
