% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{rnaCounts}
\alias{rnaCounts}
\title{DNA observations for the chromosomal-episomal dataset}
\format{integer matrix}
\source{
\url{https://www.ncbi.nlm.nih.gov/pmc/articles/PMC5204343/}
}
\usage{
rnaCounts
}
\description{
DNA observations for the chromosomal-episomal dataset
}
\keyword{datasets}
