% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_MPRAnalyze.R
\name{runMPRAnalyze}
\alias{runMPRAnalyze}
\title{Wrapper function MPRAnalyze}
\usage{
runMPRAnalyze(matRNACounts, matDNACounts, dfAnnotation, vecCtrlIDs = NULL,
  strModelFull = NULL, strModelRed = NULL, vecModelFacRNAFull = NULL,
  vecModelFacRNAFullCtrl = NULL, vecModelFacRNARed = NULL,
  vecModelFacRNARedCtrl = NULL, vecModelFacDNA = NULL, scaNProc = 1,
  vecDispersionsExternal = NULL, vecRNADepthExternal = NULL,
  vecDNADepthExternal = NULL, boolPreFitCtrlDNA = FALSE,
  strModel = "lnDNAlnRNA", boolVerbose = TRUE)
}
\arguments{
\item{matRNACounts}{(integer count matrix enhancers x samples)
RNA counts per enhancer and sample.}

\item{matDNACounts}{(integer count matrix enhancers x samples)
DNA counts per enhancer and sample.}

\item{dfAnnotation}{(data frame samples x meta data)
Contains sample names and associated batch labels. All model
factors listed below which are used to group samples into batches
must be column names in dfAnnotation.}

\item{vecCtrlIDs}{(vector of strings number of control enhancers)
[default NULL] IDs of enhancers which are scrambled/controls and
used to define a background set.}

\item{strModelFull}{(string) Not yet supported.}

\item{strModelRed}{(string) Not yet supported.}

\item{vecModelFacRNAFull}{(vector of strings lenght number of factors
in full expression model)
[Default NULL]
The model factors (columns in dfAnnotation) which are to be used 
in the full expression model. The expression model will have batch
adjustment for these factors.}

\item{vecModelFacRNAFullCtrl}{(vector of strings lenght number of 
additional factors in full expression model for control enhancers)
[Default NULL]
The model factors (columns in dfAnnotation) which are to be used 
in the full control expression model. The expression model will have batch
adjustment for these factors.}

\item{vecModelFacRNARed}{(vector of strings lenght number of factors
in reduced expression model)
[Default NULL]
The model factors (columns in dfAnnotation) which are to be used 
in the reduced expression model. The expression model will have batch
adjustment for these factors.}

\item{vecModelFacRNARedCtrl}{(vector of strings lenght number of 
additional factors in reduced expression model for control enhancers)
[Default NULL]
The model factors (columns in dfAnnotation) which are to be used 
in the reduced control expression model. The expression model will have batch
adjustment for these factors.}

\item{vecModelFacDNA}{(vector of strings lenght number of factors
in DNA model)
[Default NULL]
The model factors (columns in dfAnnotation) which are to be used 
in the DNA model. The DNA model will have batch adjustment for these factors.}

\item{scaNProc}{(scalar) [Default 1]
Number of processes to be used during parallelisation.}

\item{vecDispersionsExternal}{Not yet supported. 
External dispersion factors.}

\item{vecRNADepthExternal}{(numeric vector length number of samples)
Library depth correction factors for RNA data set, one per sample.}

\item{vecDNADepthExternal}{(numeric vector length number of samples)
Library depth correction factors for DNA data set, one per sample.}

\item{boolPreFitCtrlDNA}{(bool) [Default False]
Whether to pre-fit the control DNA models.
Can only be used if strModel=="gammaDNApoisRNA".
The size of the estimation problem is of gammaDNApoisRNA
is very large if many control sequences are given. Setting
this pre-fitting to TRUE reduces the problem size by conditioning
full and reduced model on pre-estimated DNA models of the control sequences.
This likely comes at little cost in accuracy.}

\item{strModel}{(string) [Default "lnDNAlnRNA"]
{"lnDNAlnRNA", "gammaDNApoisRNA", "gammaDNApoisRNA_coordascent"}
Model and estimation scheme to be used:
lnDNAlnRNA: DNA point estimators are pre-estimated based on a lognormal model.
The expression model is then fit based on a negative binomial likelihood of the 
RNA counts.
gammaDNApoisRNA: DNA and expression model are co-estimated, exloiting a closed
form marginalisation. The DNA counts are modeled with a beta distribution
and the RNA counts with a poisson distribution.
gammaDNApoisRNA_coordascent: As gammaDNApoisRNA but can be used to estimate
models with many control enhancers. The problem is broken up into iterative
DNA and expression model estimation until it converges to a local optimum of
the likelihood function.}

\item{boolVerbose}{(bool) [Default TRUE]
Whether to print progress statements.}
}
\value{
(MPRAnalyzeObject)
Output container class object with data frame 
with results available via list-like properties.
}
\description{
Performs model fitting and differential expression analysis
steps required within MPRAnalyze framework.
}
\details{
Note that if UMI barcodes are used to label plasmids 
with the same enhancer and these barcodes differ between enhancers, 
the UMI count of the first barcode of each plasmid 
is treated as the first sample etc.
}
\author{
David Sebastian Fischer
}
