% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelSelection.R
\name{setModel}
\alias{setModel}
\title{Set the distributional model used}
\usage{
setModel(obj, model)
}
\arguments{
\item{obj}{the MPRAnalyze object}

\item{model}{the charater identifier of the model to be used. Currently 
supported models: "ln.nb", "gamma.pois".}
}
\value{
the MPRAnalyze with the model set for the given value
}
\description{
Set the distributional model used
}
