% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulateMPRA.R
\name{simulate.mpra}
\alias{simulate.mpra}
\title{Simulate MPRA data}
\usage{
\method{simulate}{mpra}(n.case = 100, n.ctrl = 20, n.cond = 2,
  n.bc = 25, mu.dna = 100, sd.dna.cond = 2, sd.dna.bc = 0.1,
  mu.rna = 1, sd.rna = 0.2, sd.rna.cond = 1)
}
\arguments{
\item{n.case}{number of case enhancers}

\item{n.ctrl}{number of control/scrambled enhancers}

\item{n.cond}{number of conditions}

\item{n.bc}{number of barcodes observed per condition}

\item{mu.dna}{mean dna level}

\item{sd.dna.cond}{stddev to draw dna fold change 
by condition from (centred at 1)}

\item{sd.dna.bc}{stddev to draw dna fold change 
by barcode from (centred at 1)}

\item{mu.rna}{mean to draw dna fold change from}

\item{sd.rna}{stddev to draw rna slope from (centred at mu.rna)}

\item{sd.rna.cond}{stddev to draw rna slope fold change 
by condition from (centred at 1)}
}
\value{
(list)
\itemize{
\item dna dna count matrix
\item rna rna count matrix
\item colData column annotation
}
}
\description{
Simulate MPRA data
}
