% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulateMPRA.R
\name{simulateMPRA}
\alias{simulateMPRA}
\title{Simulate MPRA data}
\usage{
simulateMPRA(model = "gamma.pois", n.case = 50, n.ctrl = 50,
  frac.de = 0.5, n.cond = 2, n.bc = 25, n.reps = 2, mu.dna = 100,
  mu.sd.dna = 0.3, sd.dna = 10, sd.dna.cond = 0.2, sd.dna.bc = 0.2,
  mu.rna = 1, mu.sd.rna = 0.1, sd.rna = 0.2, sd.rna.cond = 0.1,
  sd.rna.cond_case = 0.1)
}
\arguments{
\item{model}{noise model out of {"gamma.pois", "ln.nb", "ln.ln"}}

\item{n.case}{number of case enhancers}

\item{n.ctrl}{number of control/scrambled enhancers}

\item{frac.de}{fraction of differentially active case elements}

\item{n.cond}{number of conditions}

\item{n.bc}{number of barcodes observed per condition}

\item{n.reps}{number of repeats to generate}

\item{mu.dna}{mean of distribution of offset dna level over genes}

\item{mu.sd.dna}{sd of distribution of offset dna level over genes}

\item{sd.dna}{sd of DNA distribution of each gene, dispersion if 
model is ln.nb}

\item{sd.dna.cond}{sd of normal distribution from which log barcode fold
change is drawn: exp(rnorm(1,mu=0,sd=sd.dna.cond))}

\item{sd.dna.bc}{sd of normal distribution from which log condition fold
change is drawn: exp(rnorm(1,mu=0,sd=sd.dna.cond))}

\item{mu.rna}{mean of distribution of offset rna expression over genes}

\item{mu.sd.rna}{sd of distribution of offset rna expression over genes}

\item{sd.rna}{sd of RNA distribution of each gene, ignored in gamma.pois
as variance is set by DNA distribution}

\item{sd.rna.cond}{sd of normal distribution from which log condition fold
change is drawn: exp(rnorm(1,mu=0,sd=sd.rna.cond))}

\item{sd.rna.cond_case}{sd of normal distribution from which log condition fold
change of differentially active case sequence is drawn, note this is 
multiplied with the control base-line fold change:
exp(rnorm(1,mu=0,sd=sd.rna.cond_case))}
}
\value{
(list)
\itemize{
\item dcounts.obs observed dna count matrix
\item dcounts.true underlying (no noise) dna count matrix
\item par.dna
\itemize{
\item par.dna.mu baseline dna level per element
\item par.dna.cond condition fold-change of dna level
\item par.dna.bc barcode fold-change of dna level
}
\item rcounts.obs observed rna count matrix
\item rcounts.true underlying (no noise) rna count matrix
\item par.rna
\itemize{
\item par.rna.mu baseline expression level per element
\item par.rna.cond condition fold-change of expression
\item par.rna.de.case whether or not an element is 
differentially active
}
\item colData column annotation of samples (meta data)
\item rowAnnot row annotation of enhancers (meta data)
}
}
\description{
Simulate MPRA data
}
