% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/srcMPRAnalyze_getModelFits.R
\name{summarizeModel}
\alias{summarizeModel}
\alias{getDNAFitFull}
\alias{getDNAFitRed}
\alias{getRNAFitFull}
\alias{getRNAFitRed}
\alias{getDNAModel}
\alias{getExprModelFull}
\alias{getExprModelRed}
\alias{getDNAFitFull}
\alias{getDNAFitRed}
\alias{getRNAFitFull}
\alias{getRNAFitRed}
\alias{getDNAModel}
\alias{getExprModelFull}
\alias{getExprModelRed}
\title{Return models or model fits.}
\usage{
getDNAFitFull(obj)

getDNAFitRed(obj)

getRNAFitFull(obj)

getRNAFitRed(obj)

getDNAModel(obj)

getExprModelFull(obj)

getExprModelRed(obj)
}
\arguments{
\item{obj}{(MPRAnalyzeObject) Object with fits.}
}
\value{
(matrix enhancers x samples) Matrix with models or model fits.
}
\description{
Return models or model fits.
}
\author{
David Sebastian Fischer
}
