% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hypothesisTesting.R
\name{test.coefficient}
\alias{test.coefficient}
\title{Calculate the significance of a factor in the regression model}
\usage{
test.coefficient(obj, factor, contrast)
}
\arguments{
\item{obj}{the MpraObject}

\item{factor}{the name of the factor to make the comparison on}

\item{contrast}{the character value of the factor to use as a contrast. See details.}
}
\value{
a data.frame of the results
this include the test statistic, logFC, p-value and BH-corrected FDR.
}
\description{
Calculate the significance of a factor in the regression model
}
