% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hypothesisTesting.R
\name{test.empirical}
\alias{test.empirical}
\title{test for significant activity (quantitative analysis) using various empirical
tests (see details)}
\usage{
test.empirical(obj, statistic = NULL)
}
\arguments{
\item{obj}{the MpraObject, after running an analysis function}

\item{statistic}{if null [default], the intercept term is used as the score.
An alternate score can be provided by setting 'statistic'. Must be a numeric
vector.}
}
\value{
TODO
}
\description{
test for significant activity (quantitative analysis) using various empirical
tests (see details)
}
\details{
TODO
}
