% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hypothesisTesting.R
\name{test.lrt}
\alias{test.lrt}
\title{Calculate likelihood ratio test for the specific nested model}
\usage{
test.lrt(obj)
}
\arguments{
\item{obj}{the MpraObject containing the full and reduced}

\item{dnaDesign.reduced}{the design of the DNA reduced model. If NULL, DNA is
assumed to follow same model as the full model - and raDesign.reduced must be
provided}

\item{rnaDesign.reduced}{the design of the RNA reduced model. If NULL, RNA is
assumed to follow same model as the full model, and dnaDesign.reduced must be
provided}
}
\value{
results data frame
}
\description{
Calculate likelihood ratio test for the specific nested model
}
\note{
Must be run after running fit.differential
TODO: adatapt this to only perform test and not fit reduced model
}
