% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mpraObject.R
\name{validateMpraObject}
\alias{validateMpraObject}
\title{validate an MpraObject}
\usage{
validateMpraObject(object)
}
\arguments{
\item{object}{the MpraObject instance to validate}
}
\value{
TRUE if object is valid, otherwise vector of character strings
explaining how it's invalid
}
\description{
validate an MpraObject
}
