% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysisFunctions.R
\name{analyze.lrt}
\alias{analyze.lrt}
\title{run LRT-based analysis}
\usage{
analyze.lrt(obj)
}
\arguments{
\item{obj}{the MPRAnalyze object}
}
\value{
the MPRAnalyze object, populated with fitted models
}
\description{
run LRT-based analysis
}
