% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysisFunctions.R
\name{analyze.quantitative.empirical}
\alias{analyze.quantitative.empirical}
\title{Fit the model for quantitative analysis}
\usage{
analyze.quantitative.empirical(obj, dnaDesign = NULL, rnaDesign = NULL)
}
\arguments{
\item{obj}{the MpraObject}

\item{dnaDesign}{the design of the DNA model}

\item{rnaDesign}{the design of the RNA model}
}
\value{
the MpraObject with fitted models
}
\description{
Fit the model for quantitative analysis
}
