% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objectiveFunctions.R
\name{cost.model.wctrl}
\alias{cost.model.wctrl}
\alias{cost.dna.wctrl}
\alias{cost.rna.wctrl}
\alias{cost.dna.wctrl}
\alias{cost.rna.wctrl}
\title{likelihood wrapper to compute terms of likelihood with non-zero derivative
with respect to case dna model and rna model with control enhancer observations}
\usage{
cost.dna.wctrl(theta, theta.r, llfnDNA, llfnRNA, dcounts, rcounts, log.ddepth,
  log.rdepth, ddesign.mat, rdesign.mat)

cost.rna.wctrl(theta, theta.d, theta.d.ctrl.prefit, llfnRNA, rcounts,
  log.ddepth, log.rdepth, ddesign.mat, rdesign.mat, ddesign.ctrl.mat,
  rdesign.ctrl.mat)
}
\arguments{
\item{theta}{vector of model parameters to evaluate likelihood for
(numeric, parameters)}

\item{theta.r}{vector of rna model parameters to condition likelihood on
(numeric, rna parameters)}

\item{llfnDNA}{cost function to compute dna likelihood terms on (function)}

\item{llfnRNA}{cost function to compute rna likelihood terms on (function)}

\item{dcounts}{the observed case DNA counts (integer, enhancers x samples)}

\item{rcounts}{the observed RNA counts (integer, enhancers x samples)}

\item{log.ddepth}{dna library size correction vector, log scale (numeric, samples)}

\item{log.rdepth}{rna library size correction vector, log scale (numeric, samples)}

\item{ddesign.mat}{the dna model design matrix (logical, samples x dna parameters)}

\item{rdesign.mat}{the rna model design matrix (logical, samples x rna parameters)}

\item{theta.d}{vector of dna model parameters to condition likelihood on
(numeric, dna parameters)}

\item{theta.d.ctrl.prefit}{prefit control DNA model parameters
(numeric, control enhancers x dna model parameters)}

\item{ddesign.ctrl.mat}{prefit control DNA model design
(logical, samples x dna parameters)}

\item{rdesign.ctrl.mat}{the control rna model design matrix 
(logical, samples x rna parameters)}
}
\value{
negative sum of log likelihood terms with non-zero derivative
with respect to case model
}
\description{
nomenclature cost.[model components to be estimated].wctrl
}
