% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/libraryDepth.R
\name{estimateDepthFactors}
\alias{estimateDepthFactors}
\title{estimate library size correction factors}
\usage{
estimateDepthFactors(obj, lib.factor = NULL, depthEstimator = "uq")
}
\arguments{
\item{obj}{the MpraObject}

\item{lib.factor}{the factor associating each sample to a library. Can be a
factor or the name of a column in the object's colAnnot.}

\item{depthEstimator}{a character indicating which depth estimation to use.
Currently supported values are "uq" for upper quantile (default) and "rle"
for RLE (uses geometric mean, and is therefore not recommended if libraries
have 0 counts)}
}
\value{
the MpraObject with estimated values for sequencing depth factors
}
\description{
estimate library size correction factors
}
\note{
since in most MPRA experiments multiple barcodes exist within a single
library, each column in the matrix is usually not a separate library. For this
reason, it is recommended to supply this function with the appropriate
partitioning of the data matrix columns into libraries, see lib.factor
}
