% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessFunctions.R
\name{getModelParameters.DNA}
\alias{getModelParameters.DNA}
\alias{extractModelParameters.DNA}
\alias{extractModelParameters.RNA}
\alias{getModelParameters.RNA}
\title{extract the DNA model parameters}
\usage{
getModelParameters.DNA(obj, features = NULL, full = TRUE)

getModelParameters.RNA(obj, features = NULL, full = TRUE)
}
\arguments{
\item{obj}{the MpraObject to extract the parameters from}

\item{features}{the features to extract the parameters from (by default, 
parameters will be returned for all features)}

\item{full}{if TRUE (default), return the parameters of the full model. 
Otherwise, return the parameters of the reduced model (only relevant for 
LRT-based analyses)}
}
\value{
a data.frame of features (rows) by parameters (cols). By convension, the
first parameter is related to the second moment, and the interpretation of 
it depends on the distributional model used (`alpha` for `gamma.pois`, variance 
for `ln.nb` and `ln.ln`)
}
\description{
extract the DNA model parameters
}
