% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimizers.R
\name{fit.dnarna.onlyctrl.iter}
\alias{fit.dnarna.onlyctrl.iter}
\title{fit the control-based joint model. Seperate DNA model per enhancer, joint RNA
model.}
\usage{
fit.dnarna.onlyctrl.iter(model, dcounts, rcounts, ddepth, rdepth,
  ddesign.mat, rdesign.mat, d2rdesign.mat, BPPARAM,
  print.progress = TRUE)
}
\arguments{
\item{model}{noise model}

\item{dcounts}{the DNA counts}

\item{rcounts}{the RNA counts}

\item{ddepth}{dna library size correction vector (numeric, dna samples)}

\item{rdepth}{rna library size correction vector (numeric, rna samples)}

\item{ddesign.mat}{the dna model design matrix (logical, rna samples x dna parameters)}

\item{rdesign.mat}{the rna model design matrix (logical, rna samples x rna parameters)}

\item{d2rdesign.mat}{the transition matrix relating DNA estimates to RNA observations (logical, rna sample x dna parameters)}

\item{BPPARAM}{the parallelization backend to use}

\item{print.progress}{print a progress report on the fitting (default:TRUE),
recommended since the joint model fitting is computationally intensive}
}
\description{
fit the control-based joint model. Seperate DNA model per enhancer, joint RNA
model.
}
