% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimizers.R
\name{fit.dnarna}
\alias{fit.dnarna}
\alias{fit.dnarna.noctrlobs}
\alias{fit.dnarna.wctrlobs.iter}
\alias{fit.dnarna.onlyctrl.iter}
\alias{fit.dnarna.noctrlobs}
\alias{fit.dnarna.wctrlobs.iter}
\alias{fit.dnarna.onlyctrl.iter}
\title{Fit dna and rna model to a given enhancer}
\usage{
fit.dnarna.noctrlobs(model, dcounts, rcounts, ddepth, rdepth, rctrlscale,
  ddesign.mat, rdesign.mat, rdesign.ctrl.mat, theta.d.ctrl.prefit,
  compute.hessian, BPPARAM = NULL)

fit.dnarna.wctrlobs.iter(model, dcounts, rcounts, ddepth, rdepth,
  rctrlscale = NULL, ddesign.mat, rdesign.mat, rdesign.ctrl.mat,
  theta.d.ctrl.prefit, compute.hessian, BPPARAM = NULL)

fit.dnarna.onlyctrl.iter(model, dcounts, rcounts, ddepth, rdepth,
  rctrlscale = NULL, ddesign.mat, rdesign.mat, rdesign.ctrl.mat = NULL,
  theta.d.ctrl.prefit = NULL, compute.hessian = NULL, BPPARAM)
}
\arguments{
\item{model}{noise model}

\item{dcounts}{the DNA counts}

\item{rcounts}{the RNA counts}

\item{ddepth}{dna library size correction vector (numeric, samples)}

\item{rdepth}{rna library size correction vector (numeric, samples)}

\item{ddesign.mat}{the dna model design matrix (logical, samples x dna parameters)}

\item{rdesign.mat}{the rna model design matrix (logical, samples x rna parameters)}

\item{rdesign.ctrl.mat}{the control rna model design matrix 
(logical, samples x rna parameters)}

\item{theta.d.ctrl.prefit}{ctrl dna model parameters to condition likelihood on
(numeric, control enhancers x dna parameters)}

\item{compute.hessian}{if TRUE (default), compute the Hessian matrix of the
coefficients to facilitate coefficient-based hypothesis testing}

\item{BPPARAM}{BiocParallel parallelisation parameters. Only used in 
fit.dnarna.onlyctrl.iter}
}
\value{
a list with components:
\itemize{
    \item d.fitval: the fitted values of the DNA counts
    \item d.est: the estimated true DNA levels (corrected for library size)
    \item d.coef: the fitted mode parameters for the DNA counts
    \item d.se: the standard errors of the estimates of the DNA counts
    \item r.est: the fitted values of the RNA counts
    \item r.fitval: the fitted mode parameters for the RNA counts
    \item r.est: the estimated true RNA levels (corrected for library size)
    \item r.se: the standard errors of the estimates of the RNA counts
    \item ll: the log likelihood of the model
}
}
\description{
Optim wrapper that performs numerical optimisation and extracts results.
Depending on the function chosen, a single optimisation or iterative estimation
are performed.
}
