% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilFunctions.R
\name{getDesignMat}
\alias{getDesignMat}
\title{get a design matrix from the input design}
\usage{
getDesignMat(design, annotations, condition = NULL)
}
\arguments{
\item{design}{the formula object describing the design}

\item{annotations}{the column annotations corresponding to the design}

\item{condition}{condition to substract from formula}
}
\value{
a design matrix
}
\description{
adds an extra first column if this is the rna design matrix and
the rna noise model requires a separate variance parameter which is not
used for the mean model, ie. if the first parameter to estimate is not 
part of the GLM that defines the mean parameter.
}
