% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/designMatFunctions.R
\name{getDesignMat}
\alias{getDesignMat}
\title{get a design matrix from the input design}
\usage{
getDesignMat(obj, design, testcondition = NULL)
}
\arguments{
\item{obj}{the MpraObject}

\item{design}{the input design. A matrix is returned as is, a formula is
expanded to a model matrix using the object colAnnot. If design is NULL, an
intercept-only design matrix is created and returned}

\item{testcondition}{condition to substract from formula}

\item{rna}{whether this is a rna model. Together with the noise model set
in obj@model, this defines whether an extra column for the variance link
parameter is allowed in the design matrix.}
}
\value{
a design matrix
}
\description{
adds an extra first column if this is the rna design matrix and
the rna noise model requires a separate variance parameter which is not
used for the mean model, ie. if the first parameter to estimate is not 
part of the GLM that defines the mean parameter.
}
