% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessFunctions.R
\name{getDistrParam.DNA}
\alias{getDistrParam.DNA}
\alias{getDistrParam.RNA}
\title{Get model distribution parameters from an MpraObject}
\usage{
getDistrParam.DNA(obj, enhancer = NULL, full = TRUE)

getDistrParam.RNA(obj, enhancer = NULL, full = TRUE)
}
\arguments{
\item{obj}{MpraObject to extract from}

\item{enhancer}{enhancer to extract}

\item{full}{whether to extract from full model}
}
\value{
fit parameters (numeric, samples x parameters)
}
\description{
Get model distribution parameters from an MpraObject
}
