% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessFunctions.R
\name{getFits.RNA}
\alias{getFits.RNA}
\title{Get RNA full model fits from an MpraObject}
\usage{
getFits.RNA(obj, enhancers = NULL, depth = TRUE, full = TRUE,
  rnascale = TRUE)
}
\arguments{
\item{obj}{MpraObject to extract from}

\item{enhancers}{enhancer to extract}

\item{depth}{include depth correction}

\item{full}{whether to extract from full model}

\item{rnascale}{whether to use prefit rna model factors}
}
\value{
RNA fits (numeric, enhancers x samples)
}
\description{
Get RNA full model fits from an MpraObject
}
