% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessFunctions.R
\name{getLibraryDepthFactors}
\alias{getLibraryDepthFactors}
\title{extract library depth factors.}
\usage{
getLibraryDepthFactors(obj)
}
\arguments{
\item{obj}{the object o extract the factors from}
}
\value{
a list with two elements: 'dna' with the dna library depth factors, 
and 'rna' with the rna factors
}
\description{
extract library depth factors.
}
\examples{
data <- simulateMPRA(tr = rep(2,10), da=c(rep(2,5), rep(2.5,5)), 
                     nbatch=2, nbc=20)
obj <- MpraObject(dnaCounts = data$obs.dna, 
                  rnaCounts = data$obs.rna, 
                  colAnnot = data$annot)
obj <- estimateDepthFactors(obj, lib.factor = "batch", which.lib = "both")
depth.factors <- getLibraryDepthFactors(obj)
}
