% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/likelihoodFunctions.R
\name{ll.dna}
\alias{ll.dna}
\alias{ll.dna.gamma.pois}
\alias{ll.dna.ln.nb}
\alias{ll.dna.ln.ln}
\alias{ll.dna.ln}
\title{likelihood of dna observations}
\usage{
ll.dna.gamma.pois(theta, dcounts, log.ddepth, ddesign.mat)

ll.dna.ln(theta, dcounts, log.ddepth, ddesign.mat)
}
\arguments{
\item{theta}{the vector of rna model parameters to evaluate likelihood for
(numeric, rna parameters)}

\item{dcounts}{the observed DNA counts (integer, enhancers x samples)}

\item{log.ddepth}{dna library size correction vector,
log scale (numeric, samples)}

\item{ddesign.mat}{the dna model design matrix
(logical, samples x dna parameters)}
}
\value{
negative log likelihood of dna  observations under the specified 
model
}
\description{
likelihood of dna observations
}
