% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/likelihoodFunctions.R
\name{ll.rna}
\alias{ll.rna}
\alias{ll.rna.gamma.pois}
\alias{ll.rna.ln.nb}
\alias{ll.rna.gamma.pois}
\alias{ll.rna.ln.nb}
\title{likelihood of rna observations}
\usage{
ll.rna.gamma.pois(theta, theta.d, rcounts, log.rdepth, ddesign.mat, rdesign.mat)

ll.rna.ln.nb(theta, theta.d, rcounts, log.rdepth, ddesign.mat, rdesign.mat)
}
\arguments{
\item{theta}{the vector of rna model parameters to evaluate likelihood for
(numeric, rna parameters)}

\item{theta.d}{dna model parameters to condition likelihood on}

\item{rcounts}{the observed RNA counts (integer, enhancers x samples)}

\item{log.rdepth}{rna library size correction vector, log scale (numeric, samples)}

\item{ddesign.mat}{the dna model design matrix (logical, samples x dna parameters)}

\item{rdesign.mat}{the rna model design matrix (logical, samples x rna parameters)}
}
\value{
negative log likelihood of rna observations under the specified model
}
\description{
likelihood of rna observations
}
\details{
matrix multiplications: object dimensions
theta: (rna param)
theta.d: (enhancers x dna param)
ddesign.mat: (samples x dna param-1)
rdesign.mat: (samples x rna param-1)
dcounts, rcounts, log.ddepth, log.rdepth: (samples)

matrix multiplications: log.d.est
log.d.est <- theta.d[,1] + theta.d[,-1] %*% t(ddesign.mat) (gamma.pois)
log.d.est <- theta.d[,-1] %*% t(ddesign.mat) (ln.nb)
dimensions( theta.d[,-1] %*% t(ddesign.mat) )
=(enhancers x dna param -1) x (samples x dna param-1)^T
=(enhancers x dna param -1) x (dna param-1 x samples)
=(enhancers x samples)

matrix multiplication: log.r.est
log.r.est <- log.d.est + (theta %*% t(rdesign.mat))[1,] + log.rdepth (gamma.pois)
log.r.est <- log.d.est + (theta %*% t(rdesign.mat))[1,] + log.rdepth (ln.nb)
dimensions( (theta %*% t(rdesign.mat))[1,] )
=((rna param) x (samples x rna param)^T)[1,]
=(1 x samples)[1,] = (samples)
}
