% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mpraObject.R
\name{MpraObject}
\alias{MpraObject}
\title{Initialize a MpraObject object}
\usage{
MpraObject(dnaCounts, rnaCounts, colAnnot = NULL, controls = NA_integer_,
  ncores = 1)
}
\arguments{
\item{dnaCounts}{the DNA counts matrix}

\item{rnaCounts}{the RNA counts matrix}

\item{colAnnot}{column annotations}

\item{controls}{a vector specifying which enhancers are negative controls
(scrambles)}

\item{BPPARAM}{the biocParalell backend to use for parallelization throughout
the analysis}
}
\description{
Initialize a MpraObject object
}
\examples{
##TODO
}
