% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mpraObject.R
\name{MpraObject}
\alias{MpraObject}
\title{Initialize a MpraObject object}
\usage{
MpraObject(dnaCounts, rnaCounts, dnaAnnot = NULL, rnaAnnot = NULL,
  colAnnot = NULL, controls = NA_integer_, BPPARAM = NULL)
}
\arguments{
\item{dnaCounts}{the DNA counts matrix}

\item{rnaCounts}{the RNA counts matrix}

\item{dnaAnnot}{column annotations for the DNA matrix}

\item{rnaAnnot}{column annotations for the RNA matrix}

\item{colAnnot}{(deprecated) column annotations - should only be used if DNA,
RNA annotations are the same.}

\item{controls}{a vector specifying which enhancers are negative controls
(scrambles)}

\item{BPPARAM}{the biocParalell backend to use for parallelization throughout
the analysis}
}
\value{
an MpraObject
}
\description{
Initialize a MpraObject object
}
\examples{
data <- simulateMPRA(tr = rep(2,10), da=c(rep(2,5), rep(2.5,5)), 
                     nbatch=2, nbc=20)
## use 3 of the non-active enhancers as controls
obj <- MpraObject(dnaCounts = data$obs.dna, 
                  rnaCounts = data$obs.rna, 
                  colAnnot = data$annot,
                  controls = as.integer(c(1,2,4)))
}
