% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotresults.R
\name{plotBoxplots}
\alias{plotBoxplots}
\title{Output observed ratios as boxplots and add fits}
\usage{
plotBoxplots(obj, id, condition = NULL, batch = NULL, full = TRUE)
}
\arguments{
\item{obj}{(MpraObject)
MpraObject from which enhancer is to be visualised.}

\item{id}{(string)
ID of enhancer which is to be visualised. The control set is 
taken from obj if controls were used.}

\item{condition}{(string)
Column of colAnnot that contains condition which should be resolved on
x-axis. Typically, this would be the condition that was tested for 
via analyse.condition.*()}

\item{batch}{(string)
Column of colAnnot that contains batch allocation of confounding variable
that was corrected for during fitting.}

\item{full}{(bool)
Whether to plot full or reduced model fits.}
}
\value{
(gplot) Boxplot graphic.
}
\description{
Aid intuition of model fits by visualisation of the fit 
to one enhancer.
}
