% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hypothesisTesting.R
\name{test.empirical}
\alias{test.empirical}
\title{test for significant activity (quantitative analysis) using various empirical
tests (see details)}
\usage{
test.empirical(obj, statistic = NULL)
}
\arguments{
\item{obj}{the MpraObject, after running an analysis function}

\item{statistic}{if null [default], the intercept term is used as the score.
An alternate score can be provided by setting 'statistic'. Must be a numeric
vector.}
}
\value{
a data.frame of empirical summary statistics based on the model's 
estimate of slope, or the given statistic. These are:
\itemize{
    \item statistic: the statistic (either the provided, or extracted from the
    models)
    \item zscore: Z-score of the statistic (number of standard devisations 
    from the mean)
    \item mad.score: a median-baed equivalent of the Z-score, with less 
    sensitivity to outlier values
    \item zscore.ctrl: only available if negative controls are provided.
    a Z-score based on the controls distribution, instead of the distribution 
    of the complete set of observations
    \item mad.score.ctrl: only available if negative controls are provided.
    a MAD-score based on the controls distribution, instead of the distribution 
    of the complete set of observations
    \item epval: only available if negative controls are provided. empirical 
    P-value, using the control distribution as the null
    \item fdr: only available if negative controls are provided. BH adjusted-
    empricial p-values
}
}
\description{
test for significant activity (quantitative analysis) using various empirical
tests (see details)
}
