% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hypothesisTesting.R
\name{test.lrt}
\alias{test.lrt}
\title{Calculate likelihood ratio test for the specific nested model}
\usage{
test.lrt(obj)
}
\arguments{
\item{obj}{the MpraObject containing the full and reduced}
}
\value{
results data frame
}
\description{
Calculate likelihood ratio test for the specific nested model
}
\note{
Must be run after running an LRT-based analysis
}
